agents/portfolio-assistant/requirements.txt
agent-framework>=1.2.2
agent-framework-foundry-hosting
azure-identity>=1.25.0
